refactor(app): #731 App.tsx を Project/Tabs/Team Context に分解
App.tsx (1136 行) が 5 連発の useRef ブリッジで hook 間の循環依存を
先送りする god component になっていた問題を解消する。

- lib/app-state-context.tsx (新規): useProjectLoader / useFileTabs /
  useTerminalTabs / useTeamManagement / useClaudeCheck /
  useTerminalTabsPersistence の統合層と ref ブリッジ 5 連発
  (confirmDiscardRef / projectSwitchedRef / projectLoadedRef /
  discardEditorTabsForRootRef / closeTeamRef) + fs watcher effect を
  AppStateProvider に内包。useProject() / useTabs() / useTeam() の 3
  consumer hook で必要な slice だけを公開する。
- components/AppShell.tsx (新規): 画面本体 (JSX + 画面ローカル state /
  derived / handler)。3 consumer hook で状態を購読する。
- App.tsx: Provider tree + 画面本体マウントのみ (1136 → 40 行)。
  ref ブリッジは AppStateProvider 内部へ完全に閉じ込めた。

hook 呼び出し順序・opts 配線・ref 代入位置・useMemo deps・JSX 構造は
旧 App.tsx と完全一致させた純粋リファクタで、振る舞いは変えていない。
StrictMode 二重 mount でも ref 代入は冪等・effect は cleanup 付きで安全。

Closes #731
